+++ tests/unit/data/tScience.m
@@ -73,6 +73,27 @@ classdef tScience < matlab.unittest.TestCase
             testCase.verifyTrue(ismissing(actualDerivative(1)), "First element in derivative should be missing.");
         end
 
+        % Test that "HasData" property returns "true" when data is present.
+        function hasData(testCase)
+
+            % Set up.
+            science = mag.Science(timetable(datetime("now", TimeZone = "UTC"), 1), mag.meta.Science());
+
+            % Exercise and verify.
+            testCase.verifyTrue(science.HasData, """HasData"" property should be ""true"".");
+        end
+
+        % Test that "HasData" property returns "false" when table has no
+        % data.
+        function hasData_noData(testCase)
+
+            % Set up.
+            science = mag.Science(timetable.empty(), mag.meta.Science());
+
+            % Exercise and verify.
+            testCase.verifyFalse(science.HasData, """HasData"" property should be ""false"".");
+        end
+
         % Test that "crop" method crops data based on a positive "duration"
         % object.
         function cropMethod_duration_positive(testCase)