+++ resources/ReleaseNotes.md
@@ -1,3 +1,4 @@
 # Software
 
+- Add `HasData` property to `mag.TimeSeries` and `mag,IALiRT`
 - Change all time zones to UTC